+++ srcs/mlx_game_loop/ra_raycast_helper.c
@@ -26,15 +26,15 @@ float	normalize_angle(float angle)
 	return (angle);
 }
 
-int	is_out_of_bounds(t_cub *cub, float inter_x, float inter_y)
+int	is_out_of_bounds(t_cub *cub, double inter_x, double inter_y)
 {
 	int		map_index_y;
 	int		map_index_x;
 	char	**map2d;
 
 	map2d = cub->map->map2d;
-	map_index_y = floorf(inter_y / TILE_SIZE);
-	map_index_x = floorf(inter_x / TILE_SIZE);
+	map_index_y = (int)floor(inter_y / TILE_SIZE);
+	map_index_x = (int)floor(inter_x / TILE_SIZE);
 	if (map_index_y < 0 || map_index_y >= cub->map->height)
 		return (1);
 	if (map_index_x < 0 || map_index_x >= cub->map->width)
@@ -43,15 +43,15 @@ int	is_out_of_bounds(t_cub *cub, float inter_x, float inter_y)
 		return (0);
 }
 
-int	is_wall(t_cub *cub, float inter_x, float inter_y)
+int	is_wall(t_cub *cub, double inter_x, double inter_y)
 {
 	int		map_index_y;
 	int		map_index_x;
 	char	**map2d;
 
 	map2d = cub->map->map2d;
-	map_index_y = floorf(inter_y / TILE_SIZE);
-	map_index_x = floorf(inter_x / TILE_SIZE);
+	map_index_y = (int)floor(inter_y / TILE_SIZE);
+	map_index_x = (int)floor(inter_x / TILE_SIZE);
 	if (cub->map->map2d[map_index_y]
 		&& map_index_x <= (int)ft_strlen(map2d[map_index_y]))
 		if (map2d[map_index_y][map_index_x] == '1')
@@ -59,4 +59,3 @@ int	is_wall(t_cub *cub, float inter_x, float inter_y)
 	return (0);
 	// return (cub->map->map2d[map_index_y][map_index_x] == '1');
 }
-
